+++ python/cudf/cudf/tests/test_dataframe.py
@@ -1847,6 +1847,7 @@ def gdf(pdf):
         lambda df, **kwargs: df.cumsum(**kwargs),
         lambda df, **kwargs: df.cumprod(**kwargs),
         lambda df, **kwargs: df.mean(**kwargs),
+        lambda df, **kwargs: df.median(**kwargs),
         lambda df, **kwargs: df.sum(**kwargs),
         lambda df, **kwargs: df.max(**kwargs),
         lambda df, **kwargs: df.std(ddof=1, **kwargs),
@@ -3424,8 +3425,6 @@ def test_all(data):
             assert_eq(got, expected)
         else:
             with pytest.raises(NotImplementedError):
-                gdata.all(bool_only=False)
-            with pytest.raises(NotImplementedError):
                 gdata.all(level="a")
 
     got = gdata.all()
@@ -3484,8 +3483,6 @@ def test_any(data, axis):
             expected = pdata.any(bool_only=True)
             assert_eq(got, expected)
         else:
-            with pytest.raises(NotImplementedError):
-                gdata.any(bool_only=False)
             with pytest.raises(NotImplementedError):
                 gdata.any(level="a")
